+++ v2g-liberty/CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Added
 
+- 🚀 FEAT: Residential load per phase (#471)
 - 🚀 FEAT: Warn negative grid power (#472)
 - 🚀 FEAT: Live reregister grid listeners (#473)